refactor : lottoresult의 테스트코드 추가 작성
