+++ src/main/java/lottogame/domain/dto/LottoResult.java
@@ -1,4 +1,4 @@
-package lottogame.domain.stats;
+package lottogame.domain.dto;
 
 import lottogame.domain.Rank;
